# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Booking API",
        "description": "API documentation for the Event Booking application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/bookEvent": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Book an event (User)",
                "operationId": "e6f50d543a018ed88a2c2c6728b3bbdf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id"
                                ],
                                "properties": {
                                    "event_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event booked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event booked successfully!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/BookingResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "409": {
                        "description": "You already booked this event"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/delete/{id}/event": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Cancel a booking for an event",
                "operationId": "bda2c6efa9cbce14b3e48ef5af46b915",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event to cancel booking",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Booking canceled successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Booking canceled successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Booking not found for this event."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/events": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Get all events booked by authenticated user",
                "operationId": "7efa26c7cabfaf202b1ee4d53bbf4350",
                "responses": {
                    "200": {
                        "description": "User events fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User events fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EventResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No events found for this user"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "EventResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Laravel Conference"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-06-01"
                    }
                },
                "type": "object"
            },
            "BookingResource": {
                "title": "Booking Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "event_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token with Bearer prefix",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}